+++ Cargo.toml
@@ -19,8 +19,8 @@ ftd = { version = "0.1.4", features = ["fifthtry"] }
 ft-api = { path = "ft-api", version = "0.1.0" }
 chrono-tz = "0.5"
 realm-client = { path = "realm-client", version = "0.1.0" }
-#mdbook = { git = "https://github.com/fifthtry/mdBook", version = "0.4.8"}
-mdbook = {path = "../mdBook"}
+mdbook = { git = "https://github.com/fifthtry/mdBook", version = "0.4.8", branch="fifthtry"}
+# mdbook = {path = "../mdBook"}
 
 [dependencies.chrono]
 features = ["serde"]
