+++ src/mdbook/mod.rs
@@ -1,0 +1,4 @@
+mod fenced_blocks;
+mod handle;
+
+pub use handle::handle_files;
